Separate windows and unix client_container for infra-ctl
