+++ pkg/ctl/sender/client_container_windows.go
@@ -1,0 +1,7 @@
+// Copyright 2020 New Relic Corporation. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+package sender
+
+func NewContainerClient(dockerAPIVersion, containerdNamespace, containerID, runtime string) (Client, error) {
+	return NewDockerClient(dockerAPIVersion, containerID)
+}
